requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0
sqlalchemy==2.0.25
aiosqlite==0.19.0
asyncpg==0.29.0
psycopg2-binary==2.9.9
httpx==0.26.0
neo4j==5.15.0
authlib==1.3.0
itsdangerous==2.1.2
numpy==1.26.3
openai==1.12.0
